+++ src/main/java/uz/stajirovka/jbooking/dto/response/RoomResponse.java
@@ -4,7 +4,6 @@ import uz.stajirovka.jbooking.constant.enums.Amenity;
 import uz.stajirovka.jbooking.constant.enums.MealPlan;
 import uz.stajirovka.jbooking.constant.enums.RoomType;
 
-import java.math.BigDecimal;
 import java.util.Set;
 
 public record RoomResponse(
@@ -16,7 +15,7 @@ public record RoomResponse(
         MealPlan mealPlan,
         Set<Amenity> amenities,
         Integer capacity,
-        BigDecimal pricePerNight,
+        Long pricePerNight,
         String description
 ) {
 }
