+++ src/main/java/uz/stajirovka/jbooking/service/BookingPriceService.java
@@ -2,11 +2,9 @@ package uz.stajirovka.jbooking.service;
 
 import uz.stajirovka.jbooking.entity.RoomEntity;
 
-import java.math.BigDecimal;
-
 // сервис расчёта стоимости бронирования
 public interface BookingPriceService {
 
     // рассчитывает общую стоимость бронирования
-    BigDecimal calculate(RoomEntity room, long nights);
+    Long calculate(RoomEntity room, long nights);
 }
