+++ src/main/java/uz/stajirovka/jbooking/service/BookingStatusService.java
@@ -3,20 +3,17 @@ package uz.stajirovka.jbooking.service;
 import uz.stajirovka.jbooking.dto.request.BookingModifyRequest;
 import uz.stajirovka.jbooking.dto.response.BookingResponse;
 
-import java.math.BigDecimal;
-
 // сервис управления статусами бронирования
 public interface BookingStatusService {
 
-
-    // отмена бронирования (HOLD/CONFIRMED/PAID -> CANCELLED)
+    // отмена бронирования (HOLD/CONFIRMED/MODIFIED -> CANCELLED)
     BookingResponse cancel(Long bookingId);
 
-    // изменение бронирования (CONFIRMED/PAID -> MODIFIED)
+    // изменение бронирования (HOLD/CONFIRMED/MODIFIED -> MODIFIED)
     BookingResponse modify(Long bookingId, BookingModifyRequest request);
 
     // получить сумму возврата при отмене
-    BigDecimal getRefundAmount(Long bookingId);
+    Long getRefundAmount(Long bookingId);
 
     // получить процент возврата при отмене (0-100)
     int getRefundPercent(Long bookingId);
