+++ src/main/java/uz/stajirovka/jbooking/service/CancellationPolicyService.java
@@ -2,14 +2,13 @@ package uz.stajirovka.jbooking.service;
 
 import uz.stajirovka.jbooking.entity.BookingEntity;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 // сервис политики отмены бронирования
 public interface CancellationPolicyService {
 
     // рассчитывает сумму возврата при отмене
-    BigDecimal calculateRefund(BookingEntity booking);
+    Long calculateRefund(BookingEntity booking);
 
     // возвращает процент возврата (0-100)
     int getRefundPercent(LocalDateTime checkInDate);
